requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
docker==7.1.0
docker-py==1.10.6
docker-pycreds==0.4.0
idna==3.11
requests==2.32.5
setuptools==80.9.0
six==1.17.0
urllib3==2.6.3
websocket-client==1.9.0

streamlit
anthropic
python-dotenv